Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/browsers/sandboxing/sandbox-new-execution-context-expected.txt

CONSOLE MESSAGE: Sandbox access violation: Blocked a frame at "http://localhost:8800" from accessing a frame at "http://localhost:8800".  The frame being accessed is sandboxed and lacks the "allow-same-origin" flag.
Reuse of iframe about:blank document execution context in sandbox="allow-scripts" iframe



PASS iframe with sandbox should load with new execution context
